moves point 3d
